requirements.txt
chromadb>=0.5.0
openai>=1.30.0
gradio>=4.0.0
python-dotenv>=1.0.0
